stdlib.h header package included

+++ variables.c
@@ -1,11 +1,11 @@
 #include <stdio.h>
-
+#include <stdlib.h>
 
 int main()
 {
     char stdName []= "Kofi";
     int age = 25;
-    double grade = 3.45;
+    float grade = 3.45;
 
     printf("%s is a student of UG. And he's %d years old. He has a gpa of %f\n", stdName, age, grade);
